Glue together the conversion and capture
